+++ udata_front/theme/gouvfr/__init__.py
@@ -77,7 +77,7 @@ nav.Bar('gouvfr_opendata', opendata_links)
 support_links = [
     nav.Item(_("Platform's documentation"), None, url='https://doc.data.gouv.fr'),
     nav.Item(_("Portal's API"), None, url=current_app.config.get('API_DOC_EXTERNAL_LINK', '#')),
-    nav.Item(_('Open data guides'), None, url='https://guides.etalab.gouv.fr'),
+    nav.Item(_('Open data guides'), None, url=current_app.config.get('ETALAB_GUIDES_URL', '#')),
     nav.Item(_('Contact us'), None, url='https://support.data.gouv.fr'),
 ]